+++ examples/md-flexible/src/Simulation.cpp
@@ -4,24 +4,34 @@
  * @date 01.03.2021
  */
 
-#include "Simulation.h"
+#include "autopas/molecularDynamics/LJFunctor.h"
+#include "autopas/pairwiseFunctors/FlopCounterFunctor.h"
+#include "autopas/molecularDynamics/LJFunctorAVX.h"
+#include "TypeDefinitions.h"
+#include "autopas/AutoPas.h"
 
-#include <sys/ioctl.h>
-#include <unistd.h>
+extern template class autopas::AutoPas<ParticleType>;
+extern template bool autopas::AutoPas<ParticleType>::iteratePairwise(
+    autopas::LJFunctor<ParticleType, true, true>*);
+extern template bool autopas::AutoPas<ParticleType>::iteratePairwise(
+    autopas::LJFunctor<ParticleType, true, true, autopas::FunctorN3Modes::Both, true>*);
+extern template bool autopas::AutoPas<ParticleType>::iteratePairwise(
+    autopas::LJFunctorAVX<ParticleType, true, true>*);
+extern template bool autopas::AutoPas<ParticleType>::iteratePairwise(
+    autopas::FlopCounterFunctor<ParticleType>*);
+
 #ifdef AUTOPAS_INTERNODE_TUNING
 #include <mpi.h>
 #endif
-
+#include "Simulation.h"
+#include <sys/ioctl.h>
+#include <unistd.h>
 #include <iomanip>
 #include <iostream>
-
 #include "BoundaryConditions.h"
 #include "Checkpoint.h"
 #include "Thermostat.h"
 #include "TimeDiscretization.h"
-#include "autopas/molecularDynamics/LJFunctor.h"
-#include "autopas/molecularDynamics/LJFunctorAVX.h"
-#include "autopas/pairwiseFunctors/FlopCounterFunctor.h"
 #include "autopas/utils/MemoryProfiler.h"
 
 void Simulation::initializeParticlePropertiesLibrary() {
