+++ examples/md-flexible/src/TemplateInstantiationLJFunctor1.cpp
@@ -1,0 +1,8 @@
+#include "autopas/molecularDynamics/LJFunctor.h"
+
+#include "TypeDefinitions.h"
+
+#include "autopas/AutoPasImpl.h"
+
+template bool autopas::AutoPas<ParticleType>::iteratePairwise(
+    autopas::LJFunctor<ParticleType, true, true>*);
