+++ examples/md-flexible/src/TemplateInstantiationLJFunctorAVX.cpp
@@ -1,0 +1,7 @@
+
+#include "autopas/AutoPasImpl.h"
+#include "TypeDefinitions.h"
+#include "autopas/molecularDynamics/LJFunctorAVX.h"
+
+template bool autopas::AutoPas<ParticleType>::iteratePairwise(
+    autopas::LJFunctorAVX<ParticleType, true, true>*);
