+++ examples/md-flexible/src/TemplateInstantiations.cpp
@@ -1,0 +1,6 @@
+
+#include "TypeDefinitions.h"
+
+#include "autopas/AutoPasImpl.h"
+
+template class autopas::AutoPas<ParticleType>;
